requirements.txt
Django==2.0
django-autoslug-iplweb==1.9.4.dev0
django-bootstrap4==0.0.7
django-ckeditor==5.6.1
django-crispy-forms==1.7.2
django-js-asset==1.1.0
django-model-utils==3.1.2
django-smart-selects==1.5.4
Pillow==5.3.0
progressbar2==3.38.0
psycopg2-binary==2.7.5
python-utils==2.3.0
pytz==2018.5
six==1.11.0
swapper==1.1.0
tqdm==4.27.0
Unidecode==1.0.22
